Add files via upload
Optiveriance Business Infrastructure Audit
Power BI Analysis
Profit Margin for Products

The profit margins for each product are calculated, with the total average profit margin across all products being 37.39%. The top-performing product in terms of profit margin is 'Penter' with a profit margin of 40.15%.
Top Customers

The top 7 customers in terms of sales volume were identified. 'Medline' stands as the top customer with a sales volume of 40,81,485.90. The total sales from the top 7 customers amount to 2,59,65,193.40.
Top Products

The top 7 products in terms of sales were also identified. The product 'Brimmer' generated the highest sales with a total of 2,57,10,124.40. The total sales from the top 7 products amount to 13,27,25,191.00.
Regional Sales Analysis

The total sales were 154.57M, with a total profit of 57.79M, indicating a profit margin of 37.39%. The total quantity sold was 68K.
Sales and Profit Last Year by Product

Total sales and profits from last year were calculated for each product.
Sales and Profit Last Year by Date

Data was recorded for sales and profit for each day of June 2015. The total sales for that period were 9,44,35,240.40 and the total profit was 3,53,11,846.03.
Sales in past 2 years

Sales data from two years ago (June 2016) was provided. The total sales for that period were 3,50,40,899.50.
Sales and Profit Last Year by Month

Sales and profit for the last year were calculated and compared on a month-by-month basis.
Sales and Profit by Average

Sales and profit moving averages were calculated. The sales moving average was 163.57K, and the profit moving average was 61.15K.
Profit Margin Table

The Profit Margin Table provided a quarterly analysis of total costs, total sales, total profit, and profit margin for the years 2014, 2015, and 2016. The total average profit margin for this period was 37.39%.